reorg scanner settings into separate files
what a fckn chore
